+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
-- ProofTree.count\_objects\_of\_length() implements the recurrence relation
+- `ProofTree.count_objects_of_length()` implements the recurrence relation
   implied by the proof tree as long as the strategies used are only disjoint unions,
   decompositions, verification or recursion.