+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   implied by the proof tree as long as the strategies used are only disjoint unions,
   decompositions, verification or recursion.
 
+### Removed
+- Remove the dependency on `permuta`.
+
 ## [0.2.1] - 2019-08-26
 ### Fixed
 - Update sympy version to 1.4